refactors the client PIN implementation into a new module
